+++ changes.md
@@ -3,7 +3,8 @@
 ## Update 131_cs0 (2025-01-15)
 
 * Script: Updated `checkOMTversion.groovy` to close the project after prompt to the user
-* Script: Updated `pisa25math.groovy` script to handle math letters HELP and some Math units (not just any file)
+* Script: Updated `pisa25math.groovy` script to handle math letters HELP and some Math units (not just any XML file)
+* Script: Updated `checkOMTversion.groovy` to removed support for OmegaT 5.7.2 -- time to upgrade!
 
 ## Update 130_cs0 (2024-12-13)